Fixed several circular importation
