+++ src/ExperimentTree.py
@@ -35,91 +35,13 @@ from PyQt6 import sip
 
 from ExtendGUI import LogBox, ManyButtons, PatternDialog
 from LogviewerTab import LogviewerTab
-from NewSingleTab import SingleTab, mpl_figure_scale, mpl, \
-                                FigureCanvasQTAgg, NavigationToolbar2QTAgg, plot1D
+from NewSingleTab import SingleTab
 from FileFinder import header_read
-from ADLERcalc.arrayUtils import precise_merge
+from ADLERcalc.spectrumUtils import precise_merge
 from ADLERcalc.ioUtils import read_1D_curve, simplify_number_range, resource_path
-from ADLERplot.Plotter import Plotter
 # from geom_tools import normalise, length, arb_rotation
 
 last_object = 1
-
-class MeasPointCompareDialog(QDialog):
-    values_ready = pyqtSignal(object)
-    def __init__(self, parent, dnode):
-        super().__init__(parent)
-        self.layout = QHBoxLayout(self)
-        self.rightside = QWidget(self)
-        self.rightlayout = QVBoxLayout(self.rightside)
-        mb1 = ManyButtons(self, ['FWHM', 'Area', 'Centre'], label = "Channels",  ticked = False)
-        mb2 = ManyButtons(self, ['FWHM', 'Area', 'Centre'], label = "Energy",  ticked = False)
-        mb1.new_values.connect(self.take3)
-        mb2.new_values.connect(self.take36)
-        button1 = QPushButton('Plot', self)
-        button1.clicked.connect(self.plot)
-        cb1 = QComboBox(self)
-        self.combo = cb1
-        self.rightlayout.addWidget(mb1)
-        self.rightlayout.addWidget(mb2)
-        self.rightlayout.addWidget(cb1)
-        self.rightlayout.addWidget(button1)
-        can, fig, lay = self.MakeCanvas(self)
-        self.layout.addWidget(can)
-        self.layout.addWidget(self.rightside)
-        self.canvas = can
-        self.figure = fig
-        self.plotter = Plotter(figure = self.figure)
-        self.datanode = dnode
-        self.boolvals = np.array(6*[False])
-        self.populateCbox()
-        self.setWindowTitle("Compare files in Measurement Point " + self.datanode.name)
-    @pyqtSlot(object)
-    def take3(self, nparr):
-        self.boolvals[0:3] = nparr
-    @pyqtSlot(object)
-    def take36(self, nparr):
-        self.boolvals[3:6] = nparr
-    def populateCbox(self):
-        self.combo.addItem('Node #', [])
-        vardict = self.datanode.returnXvalues()
-        names = [str(x) for x in vardict.keys()]
-        for name in names:
-            self.combo.addItem(name, vardict[name])
-        self.vardict = vardict
-    def MakeCanvas(self, parent):
-        mdpi, winch, hinch = 75, 9.0*mpl_figure_scale, 7.0*mpl_figure_scale
-        canvas = QWidget(parent)
-        layout = QVBoxLayout(canvas)
-        figure = mpl.figure(figsize = [winch, hinch], dpi=mdpi )#, frameon = False)
-        figAgg = FigureCanvasQTAgg(figure)
-        figAgg.setParent(canvas)
-        figAgg.setSizePolicy(QSizePolicy.Policy.Minimum,QSizePolicy.Policy.Minimum)
-        figAgg.updateGeometry()
-        toolbar = NavigationToolbar2QTAgg(figAgg, canvas)
-        toolbar.update()
-        layout.addWidget(figAgg)
-        layout.addWidget(toolbar)
-        return canvas, figure, layout
-    def plot(self):
-        values = self.datanode.returnYvalues()
-        xax = self.combo.currentIndex()
-        curves = []
-        if xax < 0:
-            print("Using a generic x-axis")
-            # xaxis = np.arange(len(values)) + 1
-            xlabel = "Node #"
-        else:
-            xaxis = self.combo.currentData()
-            xlabel = self.combo.currentText()
-            print("Using ", xlabel, " as x-axis")
-        if xlabel == 'Node #':
-            xaxis = np.arange(len(values)) + 1
-        for n in range(len(self.boolvals)):
-            if self.boolvals[n]:
-                curves.append(np.column_stack([xaxis, values[:, 2*n:2*n+2]]))
-        self.plotter.plot1D(curves, outFile = "", text = '', label_override = [xlabel, ""], curve_labels= [],  title = "", 
-                 autolimits = False)
 
 class MeasPointMergeDialog(QDialog):
     values_ready = pyqtSignal(object)
@@ -1472,124 +1394,6 @@ class TableView(QTableView):
             flist.append(fname)
         self.release_items.emit(flist)
 
-class View(QGraphicsView):
-    def __init__(self, scene, parent):
-        super().__init__(scene, parent)
-        self.clickedPos = None
-        self.currentItem = None
-        self._sceneRef = scene
-        self._treeRef = None
-    def resizeEvent(self, event):
-        super().resizeEvent(event)
-        self._sceneRef.setSceneRect(0, 0, self.width()-5, self.height()-5)
-    def assignTreeLayout(self, slay):
-        self._treeRef = slay
-    def contextMenuEvent(self, event):
-        menu = QMenu(self)
-        self.clickedPos = event.pos()
-        if self.clickedPos is not None:
-            try:
-                item = self._sceneRef.itemAt(self.mapToScene(self.clickedPos), QTransform.fromScale(1.0, 1.0))
-            except:
-                self.currentItem = None
-                print("Did not find an item at ",  self.clickedPos)
-            else:
-                self.currentItem = item
-                print("The item at ",  self.clickedPos,  "was", str(item))
-        self.populateMenu(menu)
-        menu.exec_(event.globalPos())
-    def populateMenu(self, menu):
-        Action = menu.addAction("Refresh")
-        Action.triggered.connect(self.refresh)
-        Action = menu.addAction("Properties")
-        Action.triggered.connect(self.editObject)
-        if self.currentItem is not None:
-            Action = menu.addAction("Delete Node")
-            Action.triggered.connect(self.removeObject)
-            if self.currentItem.innernode is None:
-                pass
-            elif self.currentItem.fixed_nest == 0:
-                Action = menu.addAction("Add Sample Holder")
-                Action.triggered.connect(self.addObject)
-            elif self.currentItem.fixed_nest == 1:
-                Action = menu.addAction("Add Sample")
-                Action.triggered.connect(self.addObject)
-            elif self.currentItem.fixed_nest == 2:
-                Action = menu.addAction("Add Measurement Point")
-                Action.triggered.connect(self.addObject)
-            elif self.currentItem.fixed_nest == 3:
-                Action = menu.addAction("Add File")
-                Action.triggered.connect(self.addObject)
-                Action2 = menu.addAction("Compare Files")
-                Action2.triggered.connect(self.runComparison)
-                Action3 = menu.addAction("Merge Files")
-                Action3.triggered.connect(self.runMerge)
-            elif self.currentItem.fixed_nest == 4:
-                Action = menu.addAction("Run ADLER!")
-                Action.triggered.connect(self.runAdler)
-            else:
-                pass
-        self.refresh()
-    def removeObject(self):
-        if self.currentItem is not None:
-            if self.currentItem.fixed_nest > 0:
-                self.currentItem.destroy()
-                self._sceneRef.removeItem(self.currentItem)
-                sip.delete(self.currentItem)
-                self.currentItem = None
-        self.refresh()
-    def addObject(self):
-        if self.currentItem is not None:
-            global last_object
-            Node(self.currentItem, [80.0, 20.0], placetag = 'Object'+str(last_object))
-        self.refresh()
-    def editObject(self):
-        if self.currentItem is not None:
-            dd = DataDialog(self, self.currentItem.innernode)
-            dd.values_ready.connect(self.currentItem.innernode.accept_values)
-            # connect some signals here...?
-            dd.show()
-            # dd.raise()
-            dd.activateWindow()
-        self.refresh()
-    def runMerge(self):
-        if self.currentItem is not None:
-            cd = MeasPointMergeDialog(self, self.currentItem.innernode) 
-            cd.show()
-            # dd.raise()
-            cd.activateWindow()
-        self.refresh()
-    def runComparison(self):
-        if self.currentItem is not None:
-            cd = MeasPointCompareDialog(self, self.currentItem.innernode) 
-            cd.show()
-            # dd.raise()
-            cd.activateWindow()
-        self.refresh()
-    def runAdler(self):
-        if self.currentItem is not None:
-            dd = AdlerDialog(self, self.currentItem.innernode)
-            dd.values_ready.connect(self.currentItem.innernode.accept_values)
-            # connect some signals here...?
-            dd.show()
-            # dd.raise()
-            dd.activateWindow()
-        self.refresh()
-    def printObject(self):
-        if self.clickedPos is not None:
-            try:
-                item = self._sceneRef.itemAt(self.clickedPos, QTransform.fromScale(1.0, 1.0))
-            except:
-                print("Did not find an item at ",  self.clickedPos)
-            else:
-                print("The item at ",  self.clickedPos,  "was", str(item))
-    def refresh(self):
-        if self._treeRef is not None:
-            self._treeRef.positionItems()
-        else:
-            self._sceneRef.update()
-    def whoami(self):
-        return 'Nodes:GraphicsView'
 
 class IconWidget(QFrame):
     def __init__(self, parent=None):
