+++ src/VariablesGUI.py
@@ -178,9 +178,9 @@ class InputVariable(QObject):
             self.maxval[0] = float(self.v_maxval)
             self.minval[0] = float(self.v_minval)
         if self.isinteger:
-            self.value = self.value.astype(np.int)
-            self.maxval = self.maxval.astype(np.int)
-            self.minval = self.minval.astype(np.int)
+            self.value = self.value.astype(int)
+            self.maxval = self.maxval.astype(int)
+            self.minval = self.minval.astype(int)
         base = QWidget(parent)
         self.normal_style = base.styleSheet()
         self.normal_font = base.font()
@@ -380,9 +380,9 @@ class IOVariable(QObject):
             self.maxval[0] = float(self.v_maxval)
             self.minval[0] = float(self.v_minval)
         if self.isinteger:
-            self.value = self.value.astype(np.int)
-            self.maxval = self.maxval.astype(np.int)
-            self.minval = self.minval.astype(np.int)
+            self.value = self.value.astype(int)
+            self.maxval = self.maxval.astype(int)
+            self.minval = self.minval.astype(int)
         base = QWidget(parent)
         label = QLabel(self.v_name, base)
         label2 = QLabel(self.v_unit, base)
